Fix device finding process
Make it so that devices matches in a search path are allocated in a
non-overlapping manner.

Make sure that if symlinks are part of a match group, the symlink is
followed and removed from the unallocated list so that it is not added
to another group where it matches directly.