Improve error reporting
Previously go func errors would be logged but did not get propagated so
there was no way to cleanup the go funcs that we'd created if one of had
failed. Maintain an internal error channel that the main go func created
by Run watches for errors from the ones that are created as a part of
start. Report errors from this internal channel to the main error
channel, errors reported to this channel will cause an ordered shutdown.